requirements.txt
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
config==0.5.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
fqdn==1.5.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
lark==1.3.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.1.4
mysql-connector-python==9.5.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.29.0
selenium==4.38.0
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.45
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==25.10.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.4
widgetsnbextension==4.0.15
wsproto==1.3.1
